[types] Move compute_pipeline_diff to feldera-types
Move compute_pipeline_diff and its helpers from the adapters crate to feldera-types so the pipeline manager can compute a pipeline diff without depending on the data plane. Narrow the inputs to PipelineConfigProgramInfo (the subset it reads) and return a new feldera-types PipelineDiffError; the controller maps it back to ControllerError.

Preliminary refactor for the diff endpoint (issue #5213).
